LayoutTests/http/tests/media/modern-media-controls/overflow-support/playback-speed-live-broadcast-expected.txt

Testing the OverflowSupport behavior for Playback Speed with a live broadcast video.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.overflow') became different from null
PASS shadowRoot.querySelector('button.overflow').getBoundingClientRect().width became different from 0
Tapping overflow button...
PASS contextmenu[0].children.filter((item) => item.title === 'Playback Speed').length is 0
PASS successfullyParsed is true

TEST COMPLETE
